put all node related stuff into the same file
